+++ src/types/webhook_info.rs
@@ -3,6 +3,8 @@ use std::net::IpAddr;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
+use crate::types::AllowedUpdate;
+
 /// Contains information about the current status of a webhook.
 ///
 /// [The official docs](https://core.telegram.org/bots/api#webhookinfo).
@@ -38,5 +40,27 @@ pub struct WebhookInfo {
 
     /// A list of update types the bot is subscribed to. Defaults to all update
     /// types.
-    pub allowed_updates: Option<Vec<String>>,
+    pub allowed_updates: Option<Vec<AllowedUpdate>>,
+}
+
+// Regression test for <https://github.com/teloxide/teloxide-core/pull/166>
+#[test]
+fn empty_url() {
+    let json = r#"{"url":"","has_custom_certificate":false,"pending_update_count":0,"allowed_updates":["message"]}"#;
+    let actual: WebhookInfo = serde_json::from_str(json).unwrap();
+    let expected = WebhookInfo {
+        url: None,
+        has_custom_certificate: false,
+        pending_update_count: 0,
+        ip_address: None,
+        last_error_date: None,
+        last_error_message: None,
+        max_connections: None,
+        allowed_updates: Some(vec![AllowedUpdate::Message]),
+    };
+
+    assert_eq!(actual, expected);
+
+    let json = r#"{"ok":true,"result":{"url":"","has_custom_certificate":false,"pending_update_count":0,"allowed_updates":["message"]}}"#;
+    serde_json::from_str::<crate::net::TelegramResponse<WebhookInfo>>(json).unwrap();
 }